move ServerRpcNew and ClientSideStatus to their own file
